+++ inst/www/templates/script_template.R
@@ -8,7 +8,7 @@ library(aNCA)
 library(dplyr)
 
 # Load raw data #
-data_path <- "../data/data.rds"
+data_path <- "../input_data.rds"
 adnca_data <- read_pk(data_path)
 
 ## Preprocess data ########################################
@@ -41,11 +41,11 @@ preprocessed_adnca <- adnca_data %>%
 
 ## Setup NCA settings in the PKNCA object ########################
 int_parameters <- settings_list$settings$int_parameters
-units_table <- settings_list$final_units
+units_table <- settings_list$units_table
 parameters_selected_per_study <- settings_list$settings$parameters$selections
 study_types_df <- settings_list$settings$parameters$types_df
 extra_vars_to_keep <-  settings_list$extra_vars_to_keep
-slope_rules <- settings_list$slope_rules$manual_slopes
+slope_rules <- settings_list$slope_rules
 
 pknca_obj <- preprocessed_adnca %>%
 
@@ -77,7 +77,12 @@ pknca_obj <- preprocessed_adnca %>%
   {
     pknca_obj <- .
     if (!is.null(units_table)) {
-      pknca_obj[["units"]] <- units_table
+      pknca_obj[["units"]] <- dplyr::rows_update(
+        pknca_obj[["units"]],
+        units_table,
+        by = c("PPTESTCD", "PPORRESU"),
+        unmatched = "ignore"
+      )
     }
     pknca_obj
   }